requirements.txt
# requirements.txt
pandas>=1.3.0
openpyxl>=3.0.0
flask>=2.0.0
python-dotenv>=0.19.0